+++ public/samples/pci-hello.js
@@ -1,0 +1,7 @@
+define(function () {
+    return {
+        hello(name) {
+            return `Hello, <strong>${name}!</strong>`;
+        }
+    };
+});